+++ moe/bandit/linkers.py
@@ -2,10 +2,11 @@
 """Links between the implementations of bandit algorithms."""
 from collections import namedtuple
 
-from moe.bandit.constant import BANDIT_EPSILON_ENDPOINT, BANDIT_UCB_ENDPOINT, EPSILON_SUBTYPE_FIRST, EPSILON_SUBTYPE_GREEDY, EPSILON_SUBTYPES, UCB_SUBTYPE_1, UCB_SUBTYPES
+from moe.bandit.constant import BANDIT_EPSILON_ENDPOINT, BANDIT_UCB_ENDPOINT, EPSILON_SUBTYPE_FIRST, EPSILON_SUBTYPE_GREEDY, EPSILON_SUBTYPES, UCB_SUBTYPE_1, UCB_SUBTYPE_1_TUNED, UCB_SUBTYPES
 from moe.bandit.epsilon_first import EpsilonFirst
 from moe.bandit.epsilon_greedy import EpsilonGreedy
 from moe.bandit.ucb1 import UCB1
+from moe.bandit.ucb1_tuned import UCB1Tuned
 
 BanditMethod = namedtuple(
         'BanditMethod',
@@ -32,6 +33,10 @@ UCB_SUBTYPES_TO_BANDIT_METHODS = {
         UCB_SUBTYPE_1: BanditMethod(
             subtype=UCB_SUBTYPE_1,
             bandit_class=UCB1,
+            ),
+        UCB_SUBTYPE_1_TUNED: BanditMethod(
+            subtype=UCB_SUBTYPE_1_TUNED,
+            bandit_class=UCB1Tuned,
             ),
         }