+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Implemented epsilon-greedy. (#255)
     * Implemented epsilon-first. (#335) 
     * Implemented UCB1. (#354)
+    * Implemented UCB1-tuned. (#366)
   * Added support for the L-BFGS-B optimizer. (#296)
   * Added GPU implementation for q,p-EI and its gradient computation. (#219)
     * Speed up GPU functions by redesign of memory allocation. (#297)